requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.7.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
tqdm>=4.65.0
seaborn>=0.12.0
tkinterdnd2>=0.3.0  # For drag and drop support
